Add verified badges to new PostCard UI
